requirements.txt
flask==2.0.1
pandas==1.3.3
openpyxl==3.0.9
werkzeug==2.0.1
Pillow==9.0.0
requests==2.27.1
numpy==1.21.2 